♻️ 抽取 formatBytes 到共享 lib 并新增 formatSpeed
